+++ src/Properties/Models/PropertyProxy.php
@@ -14,8 +14,15 @@ declare(strict_types=1);
 
 namespace Vanilo\Properties\Models;
 
+use Illuminate\Support\Collection;
 use Konekt\Concord\Proxies\ModelProxy;
+use Vanilo\Properties\Contracts\PropertyType;
 
+/**
+ * @method static PropertyType getType()
+ * @method static Collection values()
+ * @method static \Vanilo\Properties\Contracts\Property|null findBySlug(string $slug)
+ */
 class PropertyProxy extends ModelProxy
 {
 }